Change sarcadeDreive to have a double double
I don't know if this function actually does what it is supposed to do, I just made it to stop all the errors

+++ src/main/java/frc/robot/subsystems/DriveSubsystem.java
@@ -164,7 +164,7 @@ public class DriveSubsystem extends SubsystemBase {
     drive.arcadeDrive(Math.pow(xSpeed, 2), Math.pow(zRotation,2));
   }
 
-  public void sarcadeDrive (int xSpeed, double zRotation) {
+  public void sarcadeDrive (double xSpeed, double zRotation) {
     drive.arcadeDrive(xSpeed, zRotation);
   }
